refatorando taskView
