+++ js/services/DataService.js
@@ -1,0 +1,11 @@
+class DataService {
+    save(data) {
+     
+    }
+  
+    load() {
+    
+    }
+  }
+  export default DataService;
+
